Actualizacion V11

+++ BoletoBusMonolitic.Web/Data/Entities/Menu.cs
@@ -10,7 +10,7 @@ namespace BoletoBusMonolitic.Web.Data.Entities
         public int IdPlato { get; set; }
         public string? Nombre { get; set; }
         public string? Descripcion { get; set; }
-        public double Precio { get; set; }
+        public decimal Precio { get; set; }
         public string? Categoria { get; set; } 
     }
 }
