+++ BoletoBusMonolitic.Web/Data/Models/DetallePedidoModel.cs
@@ -9,7 +9,7 @@ namespace BoletoBusMonolitic.Web.Data.Models
         public int IdPedido { get; set;}
         public int IdPlato { get; set; }
         public int Cantidad { get; set; }
-        public int Subtotal { get; set; }
+        public decimal Subtotal { get; set; }
 
     }
 }
